move output component into main component as it'll be easier to access variables and it won't even be reused
